Build system update (#366)
* a bunch of updates to the build system
some minor organizational improvements for internationalization

* spacing

* formatting and filename fix

* fixing cmake for qt6

* Qt6 fixups

* a bunch of Qt6 fixups

* a few more Qt6 fixes

* make some of the santizers happy

* make regex parsing slightly more safe

* cleaning up includes and unused code slightly

* formatting

* trying to have the workflow build with multiple versions of Qt...

* ok, that looks like it worked well, let's add Qt 6.5.0 in there!

* fixing cmake for linux + qt6

* maybe this depends on Qt::Network?

* fix for slightly older Qt5

* disambiguating this call

* squash a few easily avoidable warnings

* isolating artifacts based on qt version

* more warning avoidance
updating qt installer

* some minor tweaks to improve the Qt6 build

* sigh, rolling my own printf because it was literally easier than refactoring all of the error code
to deal with warnings because C++ can't forward the information that a string is a literal between
function calls.

* working around pickyness of msvc

* Revert "sigh, rolling my own printf because it was literally easier than refactoring all of the error code"

This reverts commit bb6256d77f56ecc29808b30c97d09fe83be137c0.

* improved comments

* removing no longer needed code

* spelling

* alerting about unsupported resource compression methods

* try to force zlib compression for resources

* maybe fix deploy for qt6?

* maybe just an issue with qt 6.5.0?

* minor imrovements to the GSL stuff in cmake

* formatting

+++ src/macro.cpp
@@ -1594,7 +1594,7 @@ std::error_code lengthMS(DocumentWidget *document, Arguments arguments, DataValu
 		return ec;
 	}
 
-	*result = make_value(string.size());
+	*result = make_value(static_cast<int64_t>(string.size()));
 	return MacroErrorCode::Success;
 }
 
@@ -1989,7 +1989,7 @@ std::error_code toupperMS(DocumentWidget *document, Arguments arguments, DataVal
 		return ec;
 	}
 
-	// Allocate a new string and copy an uppercased version of the string it
+	// Allocate a new string and copy an upper-case version of the string it
 	for (char &ch : string) {
 		ch = static_cast<char>(safe_toupper(ch));
 	}
@@ -2008,7 +2008,7 @@ std::error_code tolowerMS(DocumentWidget *document, Arguments arguments, DataVal
 		return ec;
 	}
 
-	// Allocate a new string and copy an uppercased version of the string it
+	// Allocate a new string and copy an upper-case version of the string it
 	for (char &ch : string) {
 		ch = static_cast<char>(safe_tolower(ch));
 	}
